refactor: 검색 페이지 폴더 구조 변경
