+++ src/app/search/_component/RecentSearch.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import Icon from "@/app/_component/common/Icon";
 import { useState, useEffect, useCallback } from "react";
-import setSearchRecordInLocalStorage from "../utils/setSearchRecordInLocalStorage";
+import setSearchRecordInLocalStorage from "../_utils/setSearchRecordInLocalStorage";
 
 interface SearchRecordItemProps {
   itemText: string;
