+++ src/app/search/_component/SearchBar.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from "next/navigation";
 import { KeyboardEvent, useState } from "react";
 
-import setSearchRecordInLocalStorage from "../utils/setSearchRecordInLocalStorage";
+import setSearchRecordInLocalStorage from "../_utils/setSearchRecordInLocalStorage";
 import Icon from "@/app/_component/common/Icon";
 import Input from "@/app/_component/common/Input";
 import Toast from "@/app/_component/common/Toast";
