+++ src/app/search/_utils/setSearchRecordInLocalStorage.ts
